requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
boto3==1.9.126
botocore==1.12.126
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
decorator==4.4.0
Django==1.11
django-cors-headers==2.5.3
django-devserver==0.8.0
django-environ==0.4.5
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.14
drf-dynamic-fields==0.3.1
drfdocs==0.0.11
ecdsa==0.13
future==0.17.1
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
iteration-utilities==0.7.0
jedi==0.13.3
jmespath==0.9.4
line-profiler==2.1.2
Markdown==3.1
okta-jwt==1.3.5
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py-bcrypt==0.4
pyasn1==0.4.5
pycparser==2.19
pycrypto==2.6.1
Pygments==2.3.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-jose==3.0.1
pytz==2018.9
redis==3.2.1
requests==2.21.0
rsa==4.0
s3transfer==0.2.0
sentry-sdk==0.7.14
six==1.12.0
stomp.py==4.1.22
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
